Access Permissions, Security Layers added
basic security layers added anyone cannot use the application from the URL directly. if you want to create a new post then you have to be a user of this application. and if you are not an user then you can only see the posts but cannot edit and delete them.